cloud_functions/TransferPlayerToMatchServer/requirements.txt
oauth2client
google-cloud-firestore
google-api-python-client
